GPS implementation
Baked a NavMeshSurface, created a minimap camera and its UI. Adjusted the creation of spheres to only appear in walkable areas and set their Y values to above ground using ray casting. Using a line renderer created a GPS line and made it only visible in the minimap layer to the Minimap camera.